Move Endpoint and LoginManager config to example
- it allows example to be innovative and create different objects
  for a given same base class of Endpoint and LoginManager
- also fixed handling Exception with noMessage (causes null pointer
  exception in LogFormatter)
