+++ server/src/main/scala/com/lookout/borderpatrol/server/Config.scala
@@ -17,16 +17,6 @@ import io.circe.syntax._
 
 
 /**
- *  EndpointConfig
- */
-case class EndpointConfig(name: String, path: Path, hosts: Set[URL]) {
-  def toSimpleEndpoint: Endpoint = SimpleEndpoint(name, path, hosts)
-}
-object EndpointConfig {
-  def fromEndpoint(e: Endpoint): EndpointConfig = EndpointConfig(e.name, e.path, e.hosts)
-}
-
-/**
  * Where you will find the Secret Store and Session Store
  */
 object Config {
@@ -76,76 +66,6 @@ object Config {
         } yield ConsulSecretStore(key, hosts)
       case other  => Xor.left(DecodingFailure(s"Invalid secretStore: $other", c.history))
     }
-  }
-
-  /**
-   * Encoder/Decoder for LoginManager
-   *
-   * Note that Decoder for LoginManager does not work standalone, it can be only used
-   * while decoding the entire Config due to dependency issues
-   */
-  implicit val encodeBasicLoginManager: Encoder[BasicLoginManager] = Encoder.instance { blm =>
-    Json.fromFields(Seq(
-      ("name", blm.name.asJson),
-      ("type", blm.tyfe.asJson),
-      ("guid", blm.guid.asJson),
-      ("loginConfirm", blm.loginConfirm.asJson),
-      ("authorizePath", blm.authorizePath.asJson),
-      ("identityEndpoint", blm.identityEndpoint.name.asJson),
-      ("accessEndpoint", blm.accessEndpoint.name.asJson)
-    ))
-  }
-  implicit val encodeOAuth2LoginManager: Encoder[OAuth2LoginManager] = Encoder.instance { olm =>
-    Json.fromFields(Seq(
-      ("name", olm.name.asJson),
-      ("type", olm.tyfe.asJson),
-      ("guid", olm.guid.asJson),
-      ("loginConfirm", olm.loginConfirm.asJson),
-      ("identityEndpoint", olm.identityEndpoint.name.asJson),
-      ("accessEndpoint", olm.accessEndpoint.name.asJson),
-      ("authorizeEndpoint", olm.authorizeEndpoint.name.asJson),
-      ("tokenEndpoint", olm.tokenEndpoint.name.asJson),
-      ("certificateEndpoint", olm.certificateEndpoint.name.asJson),
-      ("clientId", olm.clientId.asJson),
-      ("clientSecret", olm.clientSecret.asJson)
-    ))
-  }
-  def decodeBasicLoginManager(eps: Map[String, EndpointConfig]): Decoder[BasicLoginManager] = Decoder.instance { c =>
-    for {
-      name <- c.downField("name").as[String]
-      tyfe <- c.downField("type").as[String]
-      guid <- c.downField("guid").as[String]
-      loginConfirm <- c.downField("loginConfirm").as[Path]
-      authorizePath <- c.downField("authorizePath").as[Path]
-      ieName <- c.downField("identityEndpoint").as[String]
-      ie <- Xor.fromOption(eps.get(ieName), DecodingFailure(s"identityEndpoint '$ieName' not found: ", c.history))
-      aeName <- c.downField("accessEndpoint").as[String]
-      ae <- Xor.fromOption(eps.get(aeName), DecodingFailure(s"accessEndpoint '$aeName' not found: ", c.history))
-    } yield BasicLoginManager(name, tyfe, guid, loginConfirm, authorizePath,
-      ie.toSimpleEndpoint, ae.toSimpleEndpoint)
-  }
-  def decodeOAuth2LoginManager(eps: Map[String, EndpointConfig]): Decoder[OAuth2LoginManager] = Decoder.instance { c =>
-    for {
-      name <- c.downField("name").as[String]
-      tyfe <- c.downField("type").as[String]
-      guid <- c.downField("guid").as[String]
-      loginConfirm <- c.downField("loginConfirm").as[Path]
-      ieName <- c.downField("identityEndpoint").as[String]
-      ie <- Xor.fromOption(eps.get(ieName), DecodingFailure(s"identityEndpoint '$ieName' not found: ", c.history))
-      aeName <- c.downField("accessEndpoint").as[String]
-      ae <- Xor.fromOption(eps.get(aeName), DecodingFailure(s"accessEndpoint '$aeName' not found: ", c.history))
-      auName <- c.downField("authorizeEndpoint").as[String]
-      au <- Xor.fromOption(eps.get(auName), DecodingFailure(s"authorizeEndpoint '$auName' not found: ", c.history))
-      teName <- c.downField("tokenEndpoint").as[String]
-      te <- Xor.fromOption(eps.get(teName), DecodingFailure(s"tokenEndpoint '$teName' not found: ", c.history))
-      ceName <- c.downField("certificateEndpoint").as[String]
-      ce <- Xor.fromOption(eps.get(ceName),
-        DecodingFailure(s"certificateEndpoint '$ceName' not found: ", c.history))
-      clientId <- c.downField("clientId").as[String]
-      clientSecret <- c.downField("clientSecret").as[String]
-    } yield OAuth2LoginManager(name, tyfe, guid, loginConfirm,
-      ie.toSimpleEndpoint, ae.toSimpleEndpoint, au.toSimpleEndpoint, te.toSimpleEndpoint, ce.toSimpleEndpoint,
-      clientId, clientSecret)
   }
 
   // Encoder/Decoder for ServiceIdentifier
@@ -227,22 +147,6 @@ object Config {
       case x: ConsulSecretStore => validateHostsConfig(field, "consulSecretStore", x.consulUrls)
       case _ => Set.empty[String]
     }
-  }
-
-  /**
-   * Validate Endpoint configuration
-   *
-   * @param field
-   * @param endpoints
-   * @return set of all the errors encountered during validation
-   */
-  def validateEndpointConfig(field: String, endpoints: Set[EndpointConfig]): Set[String] = {
-    // Find if endpoints have duplicate entries
-    (cond(endpoints.size > endpoints.map(m => m.name).size,
-      s"Duplicate entries for key (name) are found in the field: ${field}") ++
-
-    // Make sure hosts in Endpoint have http or https protocol
-    endpoints.map(m => validateHostsConfig(field, m.name, m.hosts)).flatten)
   }
 
   /**
